GSC-4 Cleanup and Adding Pressure
I added pressure and reconfigured the properties of the GasStateInput
to make it more conducive to working with the InputSlice. Now it reqires
two properties: an id and a display value. So id would be "temperature"
and display would be "Temperature (K)", for example.

I also did a bit of cleanup, removing log messages and commmented code.
I might still want to update the state structure before this is done.